requirements.txt
paramiko==3.4.0
requests==2.31.0
python-dotenv==1.0.1
pytest==8.0.0
setuptools==69.0.0
pyyaml==6.0.1
